requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
elevenlabs==0.2.27
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
aiofiles==23.2.1